Se refactoriza el codigo y agregan 2 nuevos endpoints para obtener materia y profesor por ID
